+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v2.9.0
+  * MuDi, Slew, Atte, Mixer2, AxBC initial releases
+  * Fix missing port information (multiple modules)
+  * Fix StereoStrip issue at very low sample rates
+
 ## v2.8.2
   * EvenVCO
     * Upsample Hard Sync and FM inputs